requirements.txt
absl-py==1.3.0
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.3
django-ckeditor==6.5.1
django-countries==7.4.2
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-rosetta==0.9.8
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dlib @ file:///C:/Users/DIMOSO%20JR/Desktop/ProjectWork/dlib-19.22.99-cp39-cp39-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==22.12.6
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mediapipe==0.9.0.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
oscrypto==1.3.0
packaging==21.3
Pillow==9.3.0
polib==1.1.1
protobuf==3.20.3
pybind11==2.10.1
pycparser==2.21
pygame==2.1.2
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF3==1.0.6
pypiwin32==223
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
twilio==7.15.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.8
